+++ foundation_utf8/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.3.0
+
+* Added `foundation.com.utf8.next_codepoint_length_unsafe/2` - This is the original `next_codepoint_length` with overlong support, it is not actively used within the module
+* Added `foundation.com.utf8.next_codepoint_scalar/2`
+
 # 1.2.0
 
 * Added `foundation.com.utf8.each_codepoint/2,3`